Removed Controller and HistoryService. Added simple line parsing of bash history file.
History will talk directly with UI, we dont really need the Controller
for now. I added a simple line parsing for the history file, next step
is to load each line into the CommandList, a StringArrayHashMap that
holds the actual commands. The key is the hash of the command trimmed of
spaces.